Reorganize directory structure as per documentation
