--- users.py
@@ -1,32 +1,0 @@
-"""Importing relevant libraries for creating the forms"""
-from flask_wtf import Form
-from wtforms import TextField, PasswordField
-from wtforms.validators import DataRequired, Length, Email, EqualTo
-
-class LoginForm(Form):
-    """Create login form"""
-    username = TextField('Username', validators=[DataRequired()])
-    password = PasswordField('Password', validators=[DataRequired()])
-
-
-class RegisterForm(Form):
-    """Create register form"""
-    username = TextField(
-        'username',
-        validators=[DataRequired(), Length(min=3, max=25)]
-    )
-    email = TextField(
-        'email',
-        validators=[DataRequired(), Email(message=None), Length(min=6, max=40)]
-    )
-    password = PasswordField(
-        'password',
-        validators=[DataRequired(), Length(min=6, max=25)]
-    )
-    confirm = PasswordField(
-        'Repeat password',
-        validators=[
-            DataRequired(), EqualTo('password', message='Passwords must match.')
-        ]
-    )
-
